@@ -1,5 +1,5 @@
 import { Component, OnInit } from "@angular/core";
-import { FormBuilder, FormGroup, Validators } from "@angular/forms";
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from "@angular/forms";
 import { Router } from "@angular/router";
 import { Store } from "@ngxs/store";
 import { VillainState } from "src/app/ngxs/states/villain.state";
@@ -20,14 +20,14 @@ import { Villain } from "../../villain.model";
   styleUrls: ["./villains.component.css"],
 })
 export class VillainsComponent implements OnInit {
-  itemForm: FormGroup;
-  editedForm: FormGroup;
+  itemForm: UntypedFormGroup;
+  editedForm: UntypedFormGroup;
   villains: Villain[];
   isLoading = false;
   editingTracker = "0";
 
   constructor(
-    private fb: FormBuilder,
+    private fb: UntypedFormBuilder,
     private router: Router,
     private store: Store
   ) {}
